--- src/main/java/com/hhp7/concert/Hhp7ConcertReservationApplication.java
@@ -1,13 +1,0 @@
-package com.hhp7.concert;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Hhp7ConcertReservationApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Hhp7ConcertReservationApplication.class, args);
-    }
-
-}
