--- src/main/java/com/hhp7/concert/concertmanagement/application/ConcertManager.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.concertmanagement.application;
-
-public class ConcertManager {
-}
