--- src/main/java/com/hhp7/concert/concertmanagement/domain/entity/ConcertSchedule.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.concertmanagement.domain.entity;
-
-public class ConcertSchedule {
-}
