--- src/main/java/com/hhp7/concert/concertmanagement/domain/repository/ConcertRepository.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.concertmanagement.domain.repository;
-
-public interface ConcertRepository {
-}
