--- src/main/java/com/hhp7/concert/concertmanagement/domain/service/ConcertService.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.concertmanagement.domain.service;
-
-public class ConcertService {
-}
