--- src/main/java/com/hhp7/concert/concertmanagement/infrastructure/ConcertJpaRepository.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.concertmanagement.infrastructure;
-
-public interface ConcertJpaRepository {
-}
