--- src/main/java/com/hhp7/concert/concertmanagement/infrastructure/ConcertRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.concertmanagement.infrastructure;
-
-public class ConcertRepositoryImpl {
-}
