--- src/main/java/com/hhp7/concert/concertmanagement/interfaces/ConcertManagementController.java
@@ -1,7 +1,0 @@
-package com.hhp7.concert.concertmanagement.interfaces;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ConcertManagementController {
-}
