--- src/main/java/com/hhp7/concert/reservationmanagement/application/ReservationManager.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationmanagement.application;
-
-public class ReservationManager {
-}
