--- src/main/java/com/hhp7/concert/reservationmanagement/domain/entity/Reservation.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationmanagement.domain.entity;
-
-public class Reservation {
-}
