--- src/main/java/com/hhp7/concert/reservationmanagement/domain/repository/ReservationRepository.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationmanagement.domain.repository;
-
-public interface ReservationRepository {
-}
