--- src/main/java/com/hhp7/concert/reservationmanagement/domain/service/ReservationService.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationmanagement.domain.service;
-
-public class ReservationService {
-}
