--- src/main/java/com/hhp7/concert/reservationmanagement/infrastructure/ReservationJpaRepository.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationmanagement.infrastructure;
-
-public interface ReservationJpaRepository {
-}
