--- src/main/java/com/hhp7/concert/reservationmanagement/infrastructure/ReservationRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationmanagement.infrastructure;
-
-public class ReservationRepositoryImpl {
-}
