--- src/main/java/com/hhp7/concert/reservationmanagement/interfaces/ReservationController.java
@@ -1,7 +1,0 @@
-package com.hhp7.concert.reservationmanagement.interfaces;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ReservationController {
-}
