--- src/main/java/com/hhp7/concert/reservationqueue/application/ReservationQueueManager.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationqueue.application;
-
-public class ReservationQueueManager {
-}
