--- src/main/java/com/hhp7/concert/reservationqueue/domain/entity/QueueToken.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationqueue.domain.entity;
-
-public class QueueToken {
-}
