--- src/main/java/com/hhp7/concert/reservationqueue/domain/repository/ReservationQueueRepository.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationqueue.domain.repository;
-
-public class ReservationQueueRepository {
-}
