--- src/main/java/com/hhp7/concert/reservationqueue/domain/service/ReservationQueueService.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationqueue.domain.service;
-
-public class ReservationQueueService {
-}
