--- src/main/java/com/hhp7/concert/reservationqueue/infrastructure/ReservationQueueJpaRepository.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationqueue.infrastructure;
-
-public interface ReservationQueueJpaRepository {
-}
