--- src/main/java/com/hhp7/concert/reservationqueue/infrastructure/ReservationQueueRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationqueue.infrastructure;
-
-public class ReservationQueueRepositoryImpl {
-}
