--- src/main/java/com/hhp7/concert/reservationqueue/interfaces/ReservationQueueController.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.reservationqueue.interfaces;
-
-public class ReservationQueueController {
-}
