--- src/main/java/com/hhp7/concert/shared/application/ConcertReservationFacade.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.shared.application;
-
-public class ConcertReservationFacade {
-}
