--- src/main/java/com/hhp7/concert/shared/scheduler/TokenInvalidationScheduler.java
@@ -1,4 +1,0 @@
-package com.hhp7.concert.shared.scheduler;
-
-public class TokenInvalidationScheduler {
-}
