+++ src/main/java/com/hhp7/concertreservation/Hhp7ConcertReservationApplication.java
@@ -1,0 +1,13 @@
+package com.hhp7.concertreservation;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Hhp7ConcertReservationApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Hhp7ConcertReservationApplication.class, args);
+    }
+
+}
