+++ src/main/java/com/hhp7/concertreservation/domain/concert/model/ConcertSchedule.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.concert.model;
+
+public class ConcertSchedule {
+}
