+++ src/main/java/com/hhp7/concertreservation/domain/concert/repository/SeatRepository.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.concert.repository;
+
+public interface SeatRepository {
+}
