+++ src/main/java/com/hhp7/concertreservation/domain/concert/service/ConsertService.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.concert.service;
+
+public class ConsertService {
+}
