+++ src/main/java/com/hhp7/concertreservation/domain/point/model/PointTransactionType.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.point.model;
+
+public enum PointTransactionType {
+}
