+++ src/main/java/com/hhp7/concertreservation/domain/point/repository/UserPointBalanceRepository.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.point.repository;
+
+public interface UserPointBalanceRepository {
+}
