+++ src/main/java/com/hhp7/concertreservation/domain/point/service/PointService.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.point.service;
+
+public class PointService {
+}
