+++ src/main/java/com/hhp7/concertreservation/domain/queue/model/Queue.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.queue.model;
+
+public class Queue {
+}
