+++ src/main/java/com/hhp7/concertreservation/domain/queue/repository/QueueRepository.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.queue.repository;
+
+public interface QueueRepository {
+}
