+++ src/main/java/com/hhp7/concertreservation/domain/queue/service/QueueService.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.queue.service;
+
+public class QueueService {
+}
