+++ src/main/java/com/hhp7/concertreservation/domain/reservation/model/Reservation.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.reservation.model;
+
+public record Reservation() {
+}
