+++ src/main/java/com/hhp7/concertreservation/domain/reservation/repository/ReservationRepository.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.reservation.repository;
+
+public interface ReservationRepository {
+}
