+++ src/main/java/com/hhp7/concertreservation/domain/reservation/service/ReservationService.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.reservation.service;
+
+public class ReservationService {
+}
