+++ src/main/java/com/hhp7/concertreservation/domain/user/model/User.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.user.model;
+
+public class User {
+}
