+++ src/main/java/com/hhp7/concertreservation/domain/user/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.user.repository;
+
+public interface UserRepository {
+}
