+++ src/main/java/com/hhp7/concertreservation/domain/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.domain.user.service;
+
+public class UserService {
+}
