+++ src/main/java/com/hhp7/concertreservation/infrastructure/persistence/jpa/entities/BaseJpaEntity.java
@@ -1,0 +1,4 @@
+package com.hhp7.concertreservation.infrastructure.persistence.jpa.entities;
+
+public class BaseJpaEntity {
+}
