+++ src/test/java/com/hhp7/concertreservation/Hhp7ConcertReservationApplicationTests.java
@@ -1,0 +1,15 @@
+package com.hhp7.concertreservation;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+
+@Import(TestcontainersConfiguration.class)
+@SpringBootTest
+class Hhp7ConcertReservationApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
