+++ src/test/java/com/hhp7/concertreservation/TestHhp7ConcertReservationApplication.java
@@ -1,0 +1,12 @@
+package com.hhp7.concertreservation;
+
+import org.springframework.boot.SpringApplication;
+
+public class TestHhp7ConcertReservationApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.from(Hhp7ConcertReservationApplication::main).with(TestcontainersConfiguration.class)
+                .run(args);
+    }
+
+}
